scripts/requirements.txt
google-generativeai
lyricsgenius
tqdm
